projects/core/gcp-oece-tech/data/legacy/repos/once-ye-s/requirements.txt
# Telegram Bot
python-telegram-bot[ext]==20.7

# Google Gemini AI
google-generativeai==0.3.2

# Redis & Database
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Async & HTTP
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.4

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Platform APIs
notion-client==2.2.1
slack-sdk==3.26.1
twilio==8.11.1
linebot==3.5.0
facebook-sdk==3.1.0
tweepy==4.14.0

# Utilities
loguru==0.7.2
schedule==1.2.0
